Switch to python 3.10
