+++ .github/workflows/lint.yml
@@ -1,0 +1,45 @@
+name: Python Checking
+
+on:
+  push:
+    branches:
+      - '**'
+    tags-ignore:
+      - '**'
+  pull_request:
+    branches:
+      - master
+
+
+jobs:
+  Lint:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.10'
+
+    - name: Pip Cache
+      uses: actions/cache@v2
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt', '**/requirements-dev.txt') }}
+        restore-keys: |
+          ${{ runner.os }}-pip-
+
+
+    - name: Install Dependencies
+      run: |
+        pip install --upgrade pip wheel
+        pip install . -r requirements-dev.txt
+
+    - name: Lint
+      run: make lint
+
+    - name: Test
+      run: make test
+
