+++ .gitlab-ci.yml
@@ -1,5 +1,4 @@
 stages:
-- lint
 - test
 - deploy
 
@@ -15,27 +14,14 @@ license_scanning:
     - if: '$CI_COMMIT_BRANCH == "master"'
     - if: '$CI_PIPELINE_SOURCE == "merge_request_event"'
 
-Lint_and_Check:
-  stage: lint
-  image: python:3.8-bullseye
-  before_script:
-    - pip install -U pip -r requirements-dev.txt --quiet
-  script:
-    - make lint
-  variables:
-    PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
-  cache:
-    paths:
-      - .cache/pip
 
-
-deploy_to_heroku:
-  image: ruby:2.7
-  stage: deploy
-  before_script:
-    - gem install dpl
-  script:
-    - dpl --provider=heroku --app=$APP_NAME --api-key=$HEROKU_API_KEY
-  rules:
-    - if: '$CI_COMMIT_TAG =~ /^v*/ && $CI_COMMIT_BRANCH == "master"'
-      when: always
+#deploy_to_heroku:
+#  image: ruby:2.7
+#  stage: deploy
+#  before_script:
+#    - gem install dpl
+#  script:
+#    - dpl --provider=heroku --app=$APP_NAME --api-key=$HEROKU_API_KEY
+#  rules:
+#    - if: '$CI_COMMIT_TAG =~ /^v*/ && $CI_COMMIT_BRANCH == "master"'
+#      when: always
