+++ Makefile
@@ -1,10 +1,10 @@
 PHONY: lint
 
 lint:
-	pylint --disable=R1710,C0209 Bot.py log_maker.py ./cogs/
-	flake8 Bot.py log_maker.py ./cogs/ --max-line-length 100 --statistics --show-source --count
-	bandit -r Bot.py log_maker.py ./cogs/
-	black --check --line-length 100 .
+	pylint --disable=R1710,C0209 ./bot
+	flake8 ./bot --max-line-length 100 --statistics --show-source --count
+	bandit -r ./bot
+	black --check --line-length 100 ./bot
 
 black-reformat:
-	black --line-length 100 .
+	black --line-length 100 ./bot
