+++ bot/cogs/graduation.py
