+++ runtime.txt
@@ -1,1 +1,1 @@
-python-3.8.12
+python-3.10.0
